mft_ui/requirements.txt
Flask==2.1.1
Werkzeug==2.2.2
python-dotenv==0.19.2